update designsystem name to design
