+++ app/build.gradle.kts
@@ -56,7 +56,7 @@ dependencies {
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
 
-    implementation(project(":core:designsystem"))
+    implementation(project(":core:design"))
 
     testImplementation(libs.junit)
