+++ core/design/.gitignore
