+++ core/design/src/main/AndroidManifest.xml
