+++ core/design/src/main/java/edu/stanford/spezikt/core/design/theme/Size.kt
@@ -1,0 +1,11 @@
+package edu.stanford.spezikt.core.design.theme
+
+import androidx.compose.ui.unit.dp
+
+// Spacing
+val SmallSpacing = 8.dp
+val MediumSpacing = 16.dp
+val LargeSpacing = 24.dp
+
+// Size
+val IconSize = 48.dp
