+++ core/design/src/main/res/drawable/ic_medication.xml
