--- core/designsystem/src/main/java/edu/stanford/spezikt/core/designsystem/theme/Color.kt
@@ -1,20 +1,0 @@
-package edu.stanford.spezikt.core.designsystem.theme
-
-import androidx.compose.ui.graphics.Color
-
-@Suppress("unused")
-val CardinalRed = Color(0xFF8C1515)
-val CardinalRedDark = Color(0xFF820000)
-val CardinalRedLight = Color(0xFFB83A4B)
-
-val White = Color(0xFFFFFFFF)
-val CoolGrey = Color(0xFF53565A)
-
-val Black = Color(0xFF2E2D29)
-val Black80 = Color(0xFF43423E)
-val Black60 = Color(0xFF767674)
-val Black40 = Color(0xFFABABA9)
-val Black20 = Color(0xFFD5D5D4)
-val Black10 = Color(0xFFEAEAEA)
-
-val RectangleBlue = Color(0xFFEBF2FC)
