--- core/designsystem/src/main/java/edu/stanford/spezikt/core/designsystem/theme/Theme.kt
@@ -1,91 +1,0 @@
-package edu.stanford.spezikt.core.designsystem.theme
-
-import android.app.Activity
-import android.os.Build
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.dynamicDarkColorScheme
-import androidx.compose.material3.dynamicLightColorScheme
-import androidx.compose.material3.lightColorScheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.WindowCompat
-
-typealias ComposableBlock = @Composable () -> Unit
-
-private val DarkColorScheme = darkColorScheme(
-    primary = CardinalRed,
-    secondary = Black60,
-    tertiary = CardinalRedDark,
-
-    background = Black,
-    surface = Black80,
-    onPrimary = White,
-    onSecondary = White,
-    onTertiary = White,
-    onBackground = White,
-    onSurface = White,
-)
-
-private val LightColorScheme = lightColorScheme(
-    primary = CardinalRed,
-    secondary = CoolGrey,
-    tertiary = RectangleBlue,
-
-    background = White,
-    surface = White,
-
-    onPrimary = White,
-    onSecondary = Black,
-    onTertiary = Black,
-
-    onBackground = Black,
-    onSurface = Black,
-)
-
-@Composable
-fun SpeziKtTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
-    // Dynamic color is available on Android 12+
-    dynamicColor: Boolean = true,
-    content: @Composable () -> Unit
-) {
-    val colorScheme = when {
-        dynamicColor && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S -> {
-            val context = LocalContext.current
-            if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
-        }
-
-        darkTheme -> DarkColorScheme
-        else -> LightColorScheme
-    }
-    val view = LocalView.current
-    if (!view.isInEditMode) {
-        SideEffect {
-            val window = (view.context as Activity).window
-            window.statusBarColor = colorScheme.primary.toArgb()
-            WindowCompat.getInsetsController(window, view).isAppearanceLightStatusBars = darkTheme
-        }
-    }
-
-    val surface: ComposableBlock = {
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = MaterialTheme.colorScheme.background,
-            content = content
-        )
-    }
-
-    MaterialTheme(
-        colorScheme = colorScheme,
-        typography = SpeziTypography,
-        content = surface
-    )
-}
